Use m + 2 instead of XSIZE.
  * replace XSIZE with m + 2
  * Fix buffer overflow.

Buffer overflow

This had a buffer overflow when pattern size was set to either pattern max and pattern max - 1, even when using XSIZE.

The pre-processing for "trans" loops in search() sets m + 2 elements (<= m + 1).  It's possible this should be (< m + 1), but it doesn't hurt it to allocate one more even if it isn't referenced.  We do need to allocate enough elements though.

Since this is a fairly obvious fix, I'll include it in the set of easy XSIZE fixes.