Add HTML stylesheet.
  * The HTML reports also require a javascript stylesheet.
  * We are putting this in the same folder as the javascript.

Old smart put the stylesheet directly in the root results folder.  This was a pain, because if you needed to zip up a report to send to someone, you had to include (a) the report (b) the javascript folder and (c) the style sheet.  It was easy to forget to include the style sheet.

Now all that is required is the report and the js folder.